Reload Nginx and HAProxy configs together and immediately after - LP:1865945
When sites are added or removed, allocated ports for Nginx cache and
backend HAproxy are reshuffled. This together with commit 8c67e60
(splay reloads) causes a race issue where there's a window where
requests hits the cache for the wrong site. Instead, we should reload
both Nginx and HAProxy at or around the same time.