Remove no-cache in HAProxy health checks
There's no need to do a full end-to-check check, which is what will be
done with no-cache. Rather, we just want to check each layer, or the
next layer HAProxy is configured for.

This is helpful for when backends are experiencing issues, we still
want the caching layer to respond with something from the cache.