Remove functional tests
They didn't work properly for a couple of reasons:

 * Since commit bf9aa16221ed3fc2daf1a98343798c8d9e428aff, the `build`
   target that `make functional` depended on is no longer present, so
   you had to build the charm manually and unpack it into
   `builds/content-cache/` in a Juju repository somewhere, or else build
   it with `charmcraft pack` and hack the functional tests to point to
   the resulting `.charm` file.

 * The tests aren't quite compatible with the current version of
   `pytest-asyncio` for some reason.

Furthermore, the integration tests are completely a superset of the
functional tests: they deal with building the charm to test themselves,
they cover the same general ground of deploying the charm using Juju to
see what happens, and they test more than the functional tests do.  So
it doesn't really make much sense to have both, especially considering
that both functional and integration tests are quite slow.