Set proxy_cache in locations without backends too
In some cases it's useful to be able to cache proxied responses even for
locations that lack a backend, for example if we're doing an internal
redirect and manually setting `proxy_pass`.  However, this is currently
difficult because the value of `proxy_cache` is generated and is
inaccessible from locations without backends.

Setting `proxy_cache` in all locations seems harmless, and allows
locations without backends to access the same cache as locations with
backends.